+++ jax/experimental/jax2tf/tests/call_tf_test.py
@@ -236,8 +236,7 @@ class CallTfTest(tf_test_util.JaxToTfTestCase):
 
   @_parameterized_jit
   def test_with_var_read(self, with_jit=True):
-    if jtu.device_under_test() == "gpu":
-      raise unittest.SkipTest("Test fails on GPU")
+    # The variable is placed on the default TF device.
     outer_var_array = np.array([3., 4.], dtype=np.float32)
     outer_var = tf.Variable(outer_var_array)
 
@@ -249,9 +248,27 @@ class CallTfTest(tf_test_util.JaxToTfTestCase):
     self.assertAllClose(x * outer_var_array + 1., res, check_dtypes=False)
 
   @_parameterized_jit
+  def test_with_var_read_other_device(self, with_jit=True):
+    # Try to place the variable in device:1.
+    outer_var_array = np.array([3., 4.], dtype=np.float32)
+
+    tf_devices = tf.config.list_logical_devices(jtu.device_under_test().upper())
+    if len(tf_devices) <= 1:
+      raise unittest.SkipTest("Only 1 TF device")
+    with tf.device(tf_devices[1]):
+      outer_var = tf.Variable(outer_var_array)
+
+    def fun_tf(x):
+      return x * outer_var + 1.
+
+    x = np.array([2., 5.,], dtype=np.float32)
+    with self.assertRaisesRegex(ValueError,
+                                r".*Error compiling TensorFlow function.*"):
+      res = _maybe_jit(with_jit, jax2tf.call_tf(fun_tf))(x)
+      self.assertAllClose(x * outer_var_array + 1., res, check_dtypes=False)
+
+  @_parameterized_jit
   def test_with_var_read_x64(self, with_jit=True):
-    if jtu.device_under_test() == "gpu":
-      raise unittest.SkipTest("Test fails on GPU")
     outer_var_array = np.array([3., 4.], dtype=np.float64)
     outer_var = tf.Variable(outer_var_array)
 
@@ -264,8 +281,6 @@ class CallTfTest(tf_test_util.JaxToTfTestCase):
 
   def test_with_var_different_shape(self):
     # See https://github.com/google/jax/issues/6050
-    if jtu.device_under_test() == "gpu":
-      raise unittest.SkipTest("Test fails on GPU")
     v = tf.Variable((4., 2.), dtype=tf.float32)
 
     def tf_func(x):
@@ -638,6 +653,7 @@ class CallTfTest(tf_test_util.JaxToTfTestCase):
     f_tf2 = jax2tf.convert(f_jax2)
     res = tf.function(f_tf2, autograph=False)(x)
     self.assertAllClose(res.numpy(), f_jax(x))
+
 
   def test_module_documentation(self):
     def cos_tf(x):