+++ CHANGELOG.md
@@ -6,6 +6,13 @@ Best viewed [here](https://jax.readthedocs.io/en/latest/changelog.html).
 Remember to align the itemized text with the first line of an item within a list.
 -->
 
+## Next jax
+* [GitHub commits](https://github.com/google/jax/compare/jax-v0.4.0...main).
+* Changes
+  * The jax2tf.call_tf function now uses for TF lowering the first TF
+    device of the same platform as used by the embedding JAX computation.
+    Before, it was using the 0th device for the JAX-default backend.
+
 ## jax 0.4.0
 * Changes
   * Support for Python 3.7 has been dropped, in accordance with JAX's